requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
anykeystore=0.2=pypi_0
apex=0.1=pypi_0
atomicwrites=1.4.1=pypi_0
attrs=22.2.0=pypi_0
blas=2.120=mkl
blas-devel=3.9.0=20_win64_mkl
boto3=1.21.32=pyhd3eb1b0_0
botocore=1.24.32=pyhd3eb1b0_0
brotlipy=0.7.0=py36h2bbff1b_1003
ca-certificates=2023.12.12=haa95532_0
certifi=2021.5.30=py36haa95532_0
cffi=1.14.6=py36h2bbff1b_0
charset-normalizer=2.0.12=pypi_0
click=8.0.3=pyhd3eb1b0_0
colorama=0.4.5=pypi_0
cryptography=35.0.0=py36h71e12ea_0
cudatoolkit=11.3.1=h280eb24_10
cxxfilt=0.3.0=pypi_0
cycler=0.11.0=pypi_0
dataclasses=0.8=pyh787bdff_2
defusedxml=0.7.1=pypi_0
docopt=0.6.2=pypi_0
easydict=1.11=pypi_0
filelock=3.4.1=pypi_0
flask=2.0.3=pypi_0
flask-cors=4.0.0=pypi_0
freetype=2.12.1=h546665d_0
greenlet=2.0.2=pypi_0
hupper=1.10.3=pypi_0
idna=3.6=pypi_0
importlib-metadata=4.8.3=pypi_0
importlib-resources=5.4.0=pypi_0
iniconfig=1.1.1=pypi_0
install=1.3.5=pypi_0
intel-openmp=2023.2.0=h57928b3_50497
itsdangerous=2.0.1=pypi_0
jinja2=3.0.3=pypi_0
jmespath=0.10.0=pyhd3eb1b0_0
joblib=1.0.1=pyhd3eb1b0_0
jpeg=9e=h8ffe710_2
kiwisolver=1.3.1=pypi_0
libblas=3.9.0=20_win64_mkl
libcblas=3.9.0=20_win64_mkl
liblapack=3.9.0=20_win64_mkl
liblapacke=3.9.0=20_win64_mkl
libpng=1.6.37=h1d00b33_4
libtiff=4.2.0=h0c97f57_3
libuv=1.44.2=h8ffe710_0
libzlib=1.2.12=h8ffe710_2
lz4-c=1.9.3=h8ffe710_1
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markupsafe=2.0.1=pypi_0
matplotlib=3.3.4=pypi_0
mkl=2023.2.0=h6a75c08_50497
mkl-devel=2023.2.0=h57928b3_50497
mkl-include=2023.2.0=h6a75c08_50497
msys2-conda-epoch=20160418=1
nltk=3.6.7=pypi_0
numpy=1.19.5=py36h4b40d73_2
oauthlib=3.2.2=pypi_0
olefile=0.46=pyh9f0ad1d_1
opencv-contrib-python=4.1.0.25=pypi_0
opencv-python=4.4.0.46=pypi_0
openssl=1.1.1w=h2bbff1b_0
packaging=21.3=pypi_0
pandas=1.1.5=pypi_0
pastedeploy=2.1.1=pypi_0
pbkdf2=1.3=pypi_0
pillow=8.1.0=py36h194931e_1
pip=21.3.1=pypi_0
plaster=1.0=pypi_0
plaster-pastedeploy=0.7=pypi_0
pluggy=1.0.0=pypi_0
prettytable=2.5.0=pypi_0
py=1.11.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.7=pypi_0
pyramid=2.0.2=pypi_0
pyramid-mailer=0.15.1=pypi_0
pysocks=1.7.1=py36haa95532_0
pytest=7.0.1=pypi_0
python=3.6.2=h09676a0_15
python-dateutil=2.8.2=pyhd3eb1b0_0
python3-openid=3.2.0=pypi_0
python_abi=3.6=2_cp36m
pytorch=1.10.1=py3.6_cuda11.3_cudnn8_0
pytorch-mutex=1.0=cuda
pytorch-pretrained-bert=0.6.2=pypi_0
pytorch-transformers=1.2.0=pypi_0
pytz=2024.1=pypi_0
pyyaml=6.0.1=pypi_0
regex=2021.8.3=py36h2bbff1b_0
repoze-sendmail=4.4.1=pypi_0
requests=2.27.1=pyhd3eb1b0_0
requests-oauthlib=1.3.1=pypi_0
s3transfer=0.5.0=pyhd3eb1b0_0
sacremoses=0.0.43=pyhd3eb1b0_0
scikit-learn=0.24.2=pypi_0
scipy=1.5.4=pypi_0
seaborn=0.11.2=pypi_0
sentencepiece=0.1.99=pypi_0
seqeval=1.2.2=pypi_0
setuptools=59.6.0=pypi_0
six=1.16.0=pyhd3eb1b0_1
sklearn=0.0=pypi_0
sqlalchemy=1.4.51=pypi_0
tbb=2021.5.0=h2d74725_1
threadpoolctl=3.1.0=pypi_0
tk=8.6.12=h8ffe710_0
tokenizers=0.7.0=pypi_0
tomli=1.2.3=pypi_0
torchaudio=0.10.1=py36_cu113
torchvision=0.11.2=py36_cu113
tqdm=4.64.1=pypi_0
transaction=3.1.0=pypi_0
transformers=2.11.0=pypi_0
translationstring=1.4=pypi_0
typing_extensions=4.1.1=pyha770c72_0
urllib3=1.26.18=pypi_0
vc=14.2=h21ff451_1
velruse=1.1.1=pypi_0
venusian=3.0.0=pypi_0
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.13=pypi_0
webob=1.8.7=pypi_0
werkzeug=2.0.3=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
win_inet_pton=1.1.0=py36haa95532_0
wincertstore=0.2=py36h7fe50ca_0
wtforms=3.0.0=pypi_0
wtforms-recaptcha=0.3.2=pypi_0
xz=5.2.6=h8d14728_0
zipp=3.6.0=pypi_0
zlib=1.2.12=h8ffe710_2
zope-deprecation=4.4.0=pypi_0
zope-interface=5.5.2=pypi_0
zope-sqlalchemy=1.6=pypi_0
zstd=1.5.2=h6255e5f_4
